Add SOCData SQL Server project to solution
Introduce a new SQL Server database project named "SOCData" to the
Visual Studio solution file `StacksOfChaos.sln`. The `SOCData.sqlproj`
file is newly added and includes configurations for building and
deploying the project in Debug and Release modes. The solution file
is updated to include the new project and its configurations.