Refactor RegisterAsync to accept JSON string input
The UserService class now includes the JsonConvert class from the Newtonsoft.Json library. The RegisterAsync method has been modified to accept a JSON string and deserialize it into a RegisterRequest object. Null checks and password matching checks have been added to ensure data integrity. The password is hashed, and a check is performed to see if the user already exists before adding the new user to the database. A new RegisterAsync method that accepts a RegisterRequest object has been added but is not yet implemented.